moved the logos to to logo folder and created glowing ellipse
