+++ src/components/programs/das/header.tsx
@@ -1,9 +1,10 @@
 import Image from "next/image";
-import Das from "@/public/das.svg";
+import Das from "@/public/logos/create.svg";
 
 const Header = () => {
   return (
-    <div className="text-acm-gray-500 relative p-10 text-2xl font-semibold">
+    <div className="text-acm-gray-500 relative overflow-hidden p-10 text-2xl font-semibold">
+      <div className="bg-acm-purple-100 absolute top-0 z-0 h-[60%] w-[30%] rounded-full opacity-50 blur-3xl" />
       <div className="text-center">
         <p className="flex justify-center p-7 text-5xl font-bold">DAS</p>
         <p className="pb-7">Data Science, AI, Statistics</p>
@@ -13,7 +14,7 @@ const Header = () => {
         skills working with Big Data, Machine Learning, and Computational Stats
         projects.{" "}
       </p>
-      <Image className="absolute top-15 right-90" src={Das} alt="Spark" />
+      <Image className="absolute top-[20%] right-[25%]" src={Das} alt="Das" />
     </div>
   );
 };
